+++ HISTORY.rst
@@ -8,6 +8,7 @@ Changelog
 - More useful ``repr`` for ``Field`` classes.
 - If a field's default is ``fields.missing`` and its serialized value is ``None``, it will not be included in the final serialized result.
 - Add ``skip_missing`` ``class Meta`` option.
+- A field's ``default`` may be a callable.
 - Allow accessor function to be configured via the ``Schema.accessor`` decorator or the ``__accessor__`` class member.
 - ``URL`` and ``Email`` fields are validated upon serialization.
 - Expanded test coverage.